// src/generated/types/EventGroupDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type EventGroupDto = {
  /**
   * @description Event group ID
   * @type string
   */
  id: string
  /**
   * @description Event ID
   * @type string
   */
  eventId: string
  /**
   * @description Event name
   * @type string
   */
  eventName: string
  /**
   * @description Lab name
   * @type string
   */
  labName: string
  /**
   * @description Date and time of the event
   * @type string, date-time
   */
  dateTime: Date
  /**
   * @description Maximum number of students
   * @type number
   */
  maxStudents: number
  /**
   * @description Number of enrolled students
   * @type number
   */
  enrolledStudents: number
  /**
   * @description Whether auto-start is enabled
   * @type boolean
   */
  autoStart: boolean
  /**
   * @description Current status of the event group
   * @type string
   */
  status: string
  /**
   * @description Actual start time
   * @type string | undefined, date-time
   */
  actualStartTime?: Date | undefined
  /**
   * @description Actual end time
   * @type string | undefined, date-time
   */
  actualEndTime?: Date | undefined
  /**
   * @description Exam mode start time
   * @type string | undefined, date-time
   */
  examModeStartTime?: Date | undefined
}