// src/generated/types/EventGroupStudentDto.ts
/**
 * Generated by Kubb (https://kubb.dev/).
 * Do not edit manually.
 */

export type EventGroupStudentDto = {
  /**
   * @description Student ID
   * @type string
   */
  studentId: string
  /**
   * @description Student name
   * @type string
   */
  studentName: string
  /**
   * @description Student username
   * @type string
   */
  username: string
  /**
   * @description Seat number
   * @type string | undefined
   */
  seatNo?: string | undefined
  /**
   * @description Whether student has attended
   * @type boolean | undefined
   */
  hasAttended?: boolean | undefined
  /**
   * @description Whether student is in exam mode
   * @type boolean
   */
  isInExamMode: boolean
  /**
   * @description When student entered exam mode
   * @type string | undefined, date-time
   */
  examModeEnteredAt?: Date | undefined
  /**
   * @description When exam started for student
   * @type string | undefined, date-time
   */
  examStartedAt?: Date | undefined
  /**
   * @description Student mark
   * @type number | undefined
   */
  mark?: number | undefined
  /**
   * @description When student submitted
   * @type string | undefined, date-time
   */
  submittedAt?: Date | undefined
  /**
   * @description Assigned exam model URL
   * @type string | undefined
   */
  assignedExamModelUrl?: string | undefined
}